src/knowledge/embedding_service.py
"""Embedding service for the knowledge management system."""

from typing import List, Union, Dict, Any
from google.cloud import aiplatform
from vertexai.generative_models import GenerativeModel

class EmbeddingService:
    """
    Service for generating embeddings for text data.
    
    This class provides methods for creating vector embeddings that can be
    used for semantic search in the knowledge management system.
    """
    
    def __init__(
        self,
        project_id: str = None,
        location: str = "us-central1",
    ):
        """
        Initialize the embedding service.
        
        Args:
            project_id: Google Cloud project ID
            location: Google Cloud region
        """
        self.project_id = project_id
        self.location = location
        self._initialized = False
        
    def initialize(self):
        """Initialize the embedding service."""
        if not self._initialized:
            # Initialize Vertex AI with project details
            aiplatform.init(project=self.project_id, location=self.location)
            self._initialized = True
    
    async def generate_embeddings(self, text: Union[str, List[str]]) -> Union[List[float], List[List[float]]]:
        """
        Generate embeddings for text using Vertex AI.
        
        Args:
            text: Text or list of texts to generate embeddings for
            
        Returns:
            Vector embedding(s) for the input text(s)
        """
        if not self._initialized:
            self.initialize()
            
        # For simplicity, using the Vertex AI Embeddings API
        # In a real implementation, you'd use the specific embedding model API
        
        # Mock implementation - in a real system, this would call the Vertex AI Embeddings API
        # This is just a placeholder that returns random vectors of the right dimension
        import random
        
        # Return a single embedding vector for a single text
        if isinstance(text, str):
            return [random.uniform(-1, 1) for _ in range(768)]
        
        # Return a list of embedding vectors for multiple texts
        return [[random.uniform(-1, 1) for _ in range(768)] for _ in text]
    
    async def generate_content_embeddings(self, content: Dict[str, Any]) -> List[float]:
        """
        Generate embeddings for structured content.
        
        Args:
            content: Dictionary containing structured content
            
        Returns:
            Vector embedding for the content
        """
        # Convert structured content to text for embedding
        if "title" in content and "description" in content:
            text = f"{content['title']}\n{content['description']}"
        elif "name" in content and "description" in content:
            text = f"{content['name']}\n{content['description']}"
        elif "text" in content:
            text = content["text"]
        else:
            # Fallback to serializing the entire dictionary
            text = str(content)
            
        # Generate and return the embedding
        return await self.generate_embeddings(text)
